refactor: move course and schedule types to interfaces
Moves the `Course` and `SavedSchedule` types to the `interfaces`
directory for better organization.
