+++ components/schedule/course-selection-table.tsx
@@ -28,7 +28,7 @@ import {
   TableRow,
 } from '@/components/ui/table';
 
-import { Course } from '@/lib/types/course';
+import { Course } from '@/lib/interfaces/course';
 import {
   formatTimeRange,
   getAvailableClasses,
