+++ lib/course-utils.ts
@@ -1,4 +1,4 @@
-import { Course } from '@/lib/types/course';
+import { Course } from '@/lib/interfaces/course';
 
 export const getFullCourseCode = (course: Course): string => {
   return `${course.code}-${course.class}`;
