+++ lib/interfaces/course.ts
