+++ lib/schedule-utils.ts
@@ -1,4 +1,4 @@
-import { Course } from '@/lib/types/course';
+import { Course } from '@/lib/interfaces/course';
 
 // Days of week array
 export const daysOfWeek = [
